WorkLogService/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
PyJWT==2.9.0
alembic==1.13.2
openai>=1.0.0
